Bunch of refactoring and restructuration
